requirements.txt
django
Pillow
djangorestframework
requests